+++ common/src/main/java/dev/tocraft/walkers/api/model/impl/SquidEntityUpdater.java
@@ -1,5 +1,6 @@
 package dev.tocraft.walkers.api.model.impl;
 
+import com.mojang.logging.LogUtils;
 import dev.tocraft.walkers.api.model.EntityUpdater;
 import dev.tocraft.walkers.mixin.accessor.SquidEntityAccessor;
 import net.fabricmc.api.EnvType;
@@ -8,43 +9,33 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.entity.animal.Squid;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.NotNull;
 
 @Environment(EnvType.CLIENT)
 public class SquidEntityUpdater<S extends Squid> implements EntityUpdater<S> {
+
     @Override
-    public void update(Player player, S squid) {
+    public void update(Player player, @NotNull S squid) {
         squid.xBodyRotO = squid.xBodyRot;
         squid.zBodyRotO = squid.zBodyRot;
         squid.oldTentacleMovement = squid.tentacleMovement;
         squid.oldTentacleAngle = squid.tentacleAngle;
-        squid.tentacleMovement += 1.0F / (squid.getRandom().nextFloat() + 1.0F) * 0.2F;
-        if ((double) squid.tentacleMovement > 6.283185307179586) {
-            squid.tentacleMovement = 6.2831855F;
-        }
-
-        if (player.isInWater()) {
-            if (squid.tentacleMovement < 3.1415927F) {
-                float f = squid.tentacleMovement / 3.1415927F;
-                squid.tentacleAngle = Mth.sin(f * f * 3.1415927F) * 3.1415927F * 0.25F;
-                if ((double) f > 0.75) {
-                    ((SquidEntityAccessor) squid).setSpeed(1.0F);
-                    ((SquidEntityAccessor) squid).setRotateSpeed(1.0F);
-                } else {
-                    ((SquidEntityAccessor) squid).setRotateSpeed(((SquidEntityAccessor) squid).getRotateSpeed() * 0.8F);
-                }
-            } else {
-                squid.tentacleAngle = 0.0F;
-                ((SquidEntityAccessor) squid).setSpeed(((SquidEntityAccessor) squid).getSpeed() * 0.9F);
-                ((SquidEntityAccessor) squid).setRotateSpeed(((SquidEntityAccessor) squid).getRotateSpeed() * 0.99F);
+        float currentTentacleSpeed = ((SquidEntityAccessor) squid).getTentacleSpeed();
+        squid.tentacleMovement += currentTentacleSpeed;
+        if (squid.tentacleMovement > Math.PI * 2) {
+            squid.tentacleMovement = (float) Math.PI * 2;
+            if (squid.getRandom().nextInt(10) == 0) {
+                ((SquidEntityAccessor) squid).setTentacleSpeed(1.0F / (squid.getRandom().nextFloat() + 1.0F) * 0.2F);
             }
+        }
+        if (player.isInWater()) {
+            Vec3 playerDeltaMovement = player.getDeltaMovement();
+            squid.yBodyRot = squid.yBodyRot + (-((float)Mth.atan2(playerDeltaMovement.x, playerDeltaMovement.z)) * (180.0F / (float)Math.PI) - squid.yBodyRot) * 0.1F;
+            squid.xBodyRot = Mth.approachDegrees(squid.xBodyRot, -player.getXRot() - 90F, 10.0F);
 
-            Vec3 vec3 = player.getDeltaMovement();
-            double d = vec3.horizontalDistance();
-            squid.zBodyRot += 3.1415927F * ((SquidEntityAccessor) squid).getRotateSpeed() * 1.5F;
-            squid.xBodyRot += (-((float) Mth.atan2(d, vec3.y)) * 57.295776F - squid.xBodyRot) * 0.1F;
         } else {
-            squid.tentacleAngle = Mth.abs(Mth.sin(squid.tentacleMovement)) * 3.1415927F * 0.25F;
-            squid.xBodyRot += (-90.0F - squid.xBodyRot) * 0.02F;
+            squid.tentacleAngle = Mth.abs(Mth.sin(squid.tentacleMovement)) * (float) Math.PI * 0.25F;
+            squid.xBodyRot = Mth.approachDegrees(squid.xBodyRot, - 90.0F, 2.0F);
         }
     }
 }